move rtdb tests to their own folder
