+++ tests/specs/rtdb/reset.spec.js
@@ -1,9 +1,9 @@
-var Rebase = require('../../dist/bundle');
+var Rebase = require('../../../dist/bundle');
 var firebase = require('firebase');
 var React = require('react');
 var ReactDOM = require('react-dom');
-var firebaseConfig = require('../fixtures/config');
-var dummyObjData = require('../fixtures/dummyObjData');
+var firebaseConfig = require('../../fixtures/config');
+var dummyObjData = require('../../fixtures/dummyObjData');
 var database = require('firebase/database');
 
 describe('reset()', function() {
