+++ tests/specs/rtdb/serverinfo.spec.js
@@ -1,8 +1,8 @@
-var Rebase = require('../../dist/bundle');
+var Rebase = require('../../../dist/bundle');
 var firebase = require('firebase');
 var database = require('firebase/database');
 
-var firebaseConfig = require('../fixtures/config');
+var firebaseConfig = require('../../fixtures/config');
 
 describe('Firebase Server Info', function() {
   var base;
